feat: standalone publish version command and variant card publish button
Extracts version publishing into a standalone PublishVersion command,
separate from environment activation. PublishToEnvironment delegates
to it. Enables publishing without environments.

Variant card shows a publish button (upload icon) when a draft exists.
On success, HX-Redirect reloads the page.
